backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.11.0
httpx==0.28.1
chromadb==1.0.15
sentence-transformers==4.1.0
openai-whisper==20250625
torch>=2.2.0
numpy>=1.26.0
pytest==8.4.1
